pyproject.toml
[tool.poetry]
name = "game-data-update"
version = "0.1.0"
description = "API & Batch code for mobile game data"
authors = ["Austin Lee <haulrest@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
polars = "^1.9.0"
oracledb = "^2.4.1"
fastapi = {extras = ["standard"], version = "^0.115.2"}
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
